requirements.txt
pyyaml>=5.0
jieba>=0.42
cryptography>=3.0
requests>=2.28
